requirement.txt
#anaconda3
Python==3.6.6 
Pytorch==1.7.0
tqdm
TensorboardX
tensorflow
matplotlib
Scipi
